requirements.txt
flask
flask-socketio
flask-cors
flask-login
eventlet
gunicorn
python-engineio
python-socketio
opencv-python-headless
numpy
onnxruntime-openvino>=1.17.1
onnx>=1.14.1
requests
psycopg2-binary